add: add Carousel component to ProgramCard
